+++ scripts/5_Mission/MetricZ/Collectors/Base.c
@@ -1,0 +1,37 @@
+/*
+    SPDX-License-Identifier: GPL-3.0-or-later
+    Copyright (c) 2025 WoozyMasta
+    Source: https://github.com/woozymasta/logz
+*/
+
+#ifdef SERVER
+/**
+    \brief Base class for MetricZ collectors.
+*/
+class MetricZ_CollectorBase
+{
+	/**
+	    \brief Component name of metrics bucket.
+	*/
+	string GetName()
+	{
+		return "unknown";
+	}
+
+	/**
+	    \brief Check if this collector is enabled.
+	*/
+	bool IsEnabled()
+	{
+		return false;
+	}
+
+	/**
+	    \brief Main flush method called by MetricZ.
+	*/
+	void Flush(MetricZ_SinkBase sink)
+	{
+		// override in subclasses
+	}
+}
+#endif