+++ CHANGELOG.md
@@ -19,8 +19,9 @@ Here is the updated changelog including the new changes.
 
 ### Added
 
-* implemented Flush State Machine, which allows processing each individual
-  block of metrics in different server frames
+* API for 3rd party mods integration `MetricZ_Exporter.Register`
+* implemented Collectors registration and framed Flush, which allows
+  processing each individual block of metrics in different server frames
 * prevent overriding reserved Prometheus labels
 * configuration has been completely moved to a new format in the JSON file
   `$profile:metricz/config.json` ⚠️